refactor(tools): split library and executable
