+++ code/hsec-tools/hsec-tools.cabal
@@ -26,7 +26,28 @@ extra-doc-files: CHANGELOG.md
 tested-with:
   GHC ==8.10.7 || ==9.0.2 || ==9.2.7 || ==9.4.5 || ==9.6.1
 
-executable hsec-render
+library
+    exposed-modules:
+                      Security.Advisories
+    build-depends:    base >=4.14 && < 4.19,
+                      text >= 1.2 && < 3,
+                      time >= 1.9 && < 1.14,
+                      Cabal >= 3.2.1.0 && < 3.11,
+                      mtl >= 2.2 && < 2.4,
+                      containers >= 0.6 && < 0.7,
+                      commonmark ^>= 0.2.2,
+                      toml-reader ^>= 0.1 || ^>= 0.2
+    hs-source-dirs:   src
+    default-language: Haskell2010
+    ghc-options:      -Wall
+                      -Wcompat
+                      -Widentities
+                      -Wincomplete-record-updates
+                      -Wincomplete-uni-patterns
+                      -Wpartial-fields
+                      -Wredundant-constraints
+
+executable hsec-tools
     main-is:          Main.hs
 
     -- Modules included in this executable, other than Main.
@@ -34,14 +55,16 @@ executable hsec-render
 
     -- LANGUAGE extensions used by modules in this package.
     -- other-extensions:
-    build-depends:    base >=4.14 && < 4.19,
+    build-depends:    hsec-tools,
+                      base >=4.14 && < 4.19,
                       text >= 1.2 && < 3,
-                      time >= 1.9 && < 1.14,
-                      Cabal >= 3.2.1.0 && < 3.11,
-                      mtl >= 2.2 && < 2.4,
-                      unordered-containers >= 0.2 && < 0.3,
-                      containers >= 0.6 && < 0.7,
-                      commonmark ^>= 0.2.2,
-                      toml-reader ^>= 0.1 || ^>= 0.2
+                      optparse-applicative == 0.17.* || == 0.18.*
     hs-source-dirs:   app
     default-language: Haskell2010
+    ghc-options:      -Wall
+                      -Wcompat
+                      -Widentities
+                      -Wincomplete-record-updates
+                      -Wincomplete-uni-patterns
+                      -Wpartial-fields
+                      -Wredundant-constraints
